+++ src/main/java/net/geforcemods/securitycraft/compat/jei/SCJEIPlugin.java
@@ -15,6 +15,7 @@ import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blocks.reinforced.IReinforcedBlock;
 import net.geforcemods.securitycraft.screen.CustomizeBlockScreen;
 import net.geforcemods.securitycraft.screen.DisguiseModuleScreen;
+import net.geforcemods.securitycraft.screen.InventoryScannerScreen;
 import net.geforcemods.securitycraft.screen.ProjectorScreen;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
@@ -66,6 +67,7 @@ public class SCJEIPlugin implements IModPlugin {
 		registration.addGuiContainerHandler(CustomizeBlockScreen.class, new SlotMover<>());
 		registration.addGuiContainerHandler(DisguiseModuleScreen.class, new SlotMover<>());
 		registration.addGuiContainerHandler(ProjectorScreen.class, new SlotMover<>());
+		registration.addGhostIngredientHandler(InventoryScannerScreen.class, new InventoryScannerGhostIngredientHandler());
 	}
 
 	@Override